requirements.txt
Flask == 3.1.0
gunicorn == 23.0.0
Scrapy == 2.12.0
tantivy == 0.22.0
PyYAML == 6.0.2
beautifulsoup4 == 4.12.3
tldextract == 5.1.3